fix(buildSrcSet): ensure operation is idempotent
The purpose of this PR is to ensure the buildSrcSet operation is
idempotent. Prior to this PR, code modified the input-parameters,
params, causing identical calls to buildSrcSet to produce different
results.

Now, the input params are copied into a new queryParams object and
this object is passed to callers requiring params. A test has been
written to show that the behavior has changed from the behavior detailed
in issue #158––the input params object remains unchanged after calling
buildSrcSet and that calling buildSrcSet multiple times produces
the same result (given the same inputs).

I also ripgrep'd through the repo with `rg 'params.[:alpha:]'` and with
`rg 'params\['` to ensure params is never mutated (i.e. it never
appears on the left hand side of any expression).

Closes #158