fix(buildSrcSet): improve spacing between widths
With some basic maths we can calculate an adjusted increment ratio from
the width tolerance which leads to a sequence of widths which remains
regular(ly exponential) right up to and including the largest width.

This avoids the case often seen where the largest two sizes are
significantly closer-spaced than any other pair.