feat: use mjs file extensions with type module (#209)
The purpose of this PR is to simplify testing es6 module code in a
node environment (travis ci). Prior to this change, mocha and node
expected module code because that's what was declared in `package.json`.
Now, `.mjs` extensions are used explicitly.

Using the `.mjs` extension tells node and mocha that they should expect
to handle modules rather than `.cjs` code. At this time I think module
support in node is still considered experimental and including the `.mjs`
extensions ensures this support is "turned on."