feat: create a DPR srcset when a fixed height is specified (#215)
The purpose of this PR is to ensure we create a DPR-based srcset when
a fixed height, h, is specified. Prior, we required that a h and
an aspect ratio (ar) be passed. Now, we require that only a w or
h be specified to trigger a DPR-based srcset to be created.

Closes #177