refactor: improve target widths generation
Prior, this function performed an array access, property access, and
a comparison in the loop condition. Now, this has been reduced to a
single comparison. Prior, this function made an additional function
call/comparison in the body (i.e. `Math.min`) on every iteration. Now,
that has been removed as well.

Prior, this function attempted to return a list resolutions before
checking the cache (i.e. `[minW]`). Now, we check the cache first.

Now, when the loop terminates we know that the last width to be pushed
is less than the currentWidth that caused the loop condition to fail
and its value is greater than or equal to `maxW`. We want to stop at
`maxW` exactly so we push it if it's less than the last resolution.