feat: use explicit extensions for main, module, and browser (#225)
The purpose of this PR is to ensure consumability for js (browser),
cjs (node), and mjs (es6 module) users. Prior, we used js extensions
for our distributables, but that caused issues when trying to consume
the various dists. Now, the file extensions have been made explicit
both in the package.json (and therefore in `dist/`). By doing this
we should be able to serve the needs of umd, esm, and cjs users.