feat: create extractUrl helper
`extractUrl()` returns URL components from a source URL string. It does
this by matching the URL against regular expressions. The irrelevant
(entire URL) matches are removed and the rest stored as their
corresponding URL components in a POJO. The function uses the ufo module
to handle most of the parsing logic. The only custom behavior is using
the `useHttps` boolean to prepend a scheme to the source URL.

- `url` can be a partial, full URL, or full proxy URL.
- `useHttps` boolean defaults to false.

The function returns `{ protocol, auth, host, pathname, search, hash }`
POJO extracted from the URL.