refactor: access settings through setters and getters
BREAKING CHANGE: this commit refactors `this.settings` so that it has a
dedicated getter `get settings()` and accessing its properties is now
done through setters and getters. IE, to get `libraryParam` you'd still
do `this.libraryParam` but under the hood, we defined a `get` and a
`set` `libraryParam`.

This fixes and issue where the object referenced would change between
calls, resulting in bugs like not being able to turn of the ixlib
params. This would happen because certain parts of the code called
`this.settings` while others called the properties directly. By
refactoring so that they all get/set settings, we avoid this issue.